+++ client/app/admin/components/CaseflowDistribution/InteractableLevers.module.scss
@@ -103,6 +103,10 @@
 
 .docketLeverRight {
   vertical-align: middle;
+  button {
+    vertical-align: middle;
+    display: inline-block;
+  }
 }
 
 .leverDisabled {
@@ -120,4 +124,14 @@
       margin-left: 28px;
       margin-bottom: 10px;
   }
+}
+
+.toggleInputHide {
+    display: none;
+}
+
+.toggleSwichInput {
+    width: auto;
+    display: inline-block;
+    vertical-align: middle;
 }
