+++ client/app/admin/components/CaseflowDistribution/InteractableLeversWrapper.jsx
@@ -1,25 +1,42 @@
 // client/app/admin/components/CaseflowDistribution/InteractableLeverWrapper.js
 
 import React from 'react';
-import InteractableLever from './InteractableLever';
 import PropTypes from 'prop-types';
+import BatchSize from './BatchSize';
+import DocketTimeGoals from './DocketTimeGoals';
+import AffinityDays from './AffinityDays';
+import styles from './InteractableLevers.module.scss';
 
-const InteractableLeverWrapper = ({ InteractableLevers, levers }) => {
+const InteractableLeverWrapper = ({ levers }) => {
+
+  const batchSizeLevers = levers.filter((lever) => {
+    return lever.data_type === "number";
+  });
+
+  const docketLevers = levers.filter((lever) => {
+    return lever.data_type === "combination";
+  });
+
+  const affinityLevers = levers.filter((lever) => {
+    return lever.data_type === "radio";
+  });
+
   return (
-    <div>
-      <h3>Active Levers</h3>
-      {InteractableLevers.map((leverItem) => {
-        const lever = levers.find((l) => l.item === leverItem);
-        return lever ? (
-          <InteractableLever key={lever.item} lever={lever} />
-        ) : null;
-      })}
+    <div className={styles.leverContainer}>
+      <div className={styles.leverContent}>
+        <div className={styles.leverHead}>
+        <div className={styles.leverH2}>Active Data Elements</div>
+          <p>You may make changes to the Case Distribution algorithm values based on the data elements below. Changes will be applied to the next scheduled case distribution event unless subsequent confirmed changes are made to the same variable.</p>
+        </div>
+      </div>
+      <BatchSize batchSizeLevers={batchSizeLevers} />
+      <AffinityDays affinityLevers={affinityLevers} />
+      <DocketTimeGoals docketLevers={docketLevers} />
     </div>
   );
 };
 
 InteractableLeverWrapper.propTypes = {
-  InteractableLevers: PropTypes.array.isRequired,
   levers: PropTypes.array.isRequired,
 };
